bg(auth): fix issues with imports issues
