+++ package.json
@@ -57,7 +57,6 @@
     "dotenv": "^7.0.0",
     "express": "^4.16.4",
     "husky": "^1.3.1",
-    "joi": "^14.3.1",
     "jsonwebtoken": "^8.5.1",
     "joi": "^14.3.1",
     "lint-staged": "^8.1.5",
