+++ server/controllers/index.js
@@ -1,9 +1,11 @@
 import profileController from './profile.controllers';
 import articleController from './article.controllers';
 import authController from './auth.controller';
+import ResetPasswordController from './resetpassword';
 
 export default {
   profileController,
   articleController,
   authController,
+  ResetPasswordController,
 };
