+++ server/routes/index.js
@@ -1,7 +1,7 @@
 import express from 'express';
 
 import getArticles from './get-articles.routes';
-import resetPasswordRouter from './resetpassword';
+import resetPasswordRouter from './resetpassword.routes';
 import profileRoute from './profile.routes';
 import articleRoute from './article.routes';
 import authRoute from './user.routes';
