+++ server/routes/user.routes.js
@@ -2,18 +2,12 @@ import express from 'express';
 import controllers from '../controllers';
 import middlewares from '../middlewares';
 
-const { signUpValidator } = middlewares;
+const { signUpValidator: auth } = middlewares;
 
 const authRoute = express.Router();
 const { authController } = controllers;
 
-authRoute.post('/signup', signUpValidator.signUpValidator);
-authRoute.post(
-  '/login',
-  signUpValidator.loginValidator,
-  authController.loginController
-);
-authRoute.post('/signup', signUpValidator.signUpValidator);
-authRoute.post('/login', signUpValidator.loginValidator);
+authRoute.post('/signup', auth.signUpValidator);
+authRoute.post('/login', auth.loginValidator, authController.loginController);
 
 export default authRoute;
